Improve BearerToken validation and UUID.toString performance (#68)
It takes 3600 nanoseconds to parse a bearer token, when the code
is already cached etc. This affects every method call. Given that this
is common library code, I'd suggest that it's worth optimising the
validation check, because this method will be called so many times.

If for example you have 10k requests per core, you are spending 7% of
your time parsing bearer tokens.

This PR changes the mechanism to be something a bit more hand coded
which obeys the same semantics, and doing so achieves 480 nanoseconds,
which takes you down to about 1%.